requirements.txt
fastapi
uvicorn
streamlit
qdrant-client
openai
python-dotenv
requests
beautifulsoup4
langchain-text-splitters
tiktoken
pypdf
rich